Move ``ensure`` tests to a file of their own.
